Create a system for generating second level Menu segment paths to avoid "one fact in two places"
