+++ src/app/core/main-menu/main-menu.component.ts
@@ -2,8 +2,8 @@ import { Component, Input } from '@angular/core';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
+import { SegmentPaths } from 'src/app/segment-path';
 import { MenuItem } from '../../models/menu-item.model';
-import { SegmentPaths } from '../../segment-paths';
 
 @Component({
   selector: 'app-main-menu',
