--- src/app/segment-paths.ts
@@ -1,43 +1,0 @@
-export class SegmentPaths {
-
-  /*
-   * Main menu
-   */
-
-  static developer = 'developer';
-  static designer = 'designer';
-
-
-  /*
-   * Library top level menu
-   */
-
-  static visualIdentity = 'visual-identity';
-  static components = 'components';
-  static modules = 'modules';
-
-
-  /*
-   * Library second level menus
-   */
-
-  // Visual identity
-  static color = 'color';
-  static icons = 'icons';
-  static typography = 'typography';
-
-  // Components
-  static accordion = 'accordion';
-  static alerts = 'alerts';
-  static badge = 'badge';
-  static buttons = 'buttons';
-  static forms = 'forms';
-  static modal = 'modal';
-  static pagination = 'pagination';
-  static search = 'search';
-  static table = 'table';
-  static tags = 'tags';
-  static toast = 'toast';
-  static tooltip = 'tooltip';
-
-}
